ref: Move HubImpl to its own file
Avoids a lot of annoying `#[cfg]` wrapping and should make it easier to add things.
